Refactor type coercion functions in datafusion/expr module
